+++ tests/workflow/test_workflow_from_config.py
@@ -354,6 +354,8 @@ class TestInterpretPreprocessors:
     (detector_entry, anomaly_detection.MatrixProfileDetector, {'window_size': 25, 'normalize': True, 'p': 1.5, 'k': 5}),
     (detector_entry, anomaly_detection.MedianMethod, {'neighborhood_size_before': 15}),
     (detector_entry, anomaly_detection.MedianMethod, {'neighborhood_size_before': 25, 'neighborhood_size_after': 5}),
+    (detector_entry, anomaly_detection.OneClassSupportVectorMachine, {'window_size': 15}),
+    (detector_entry, anomaly_detection.OneClassSupportVectorMachine, {'window_size': 15, 'kernel': 'poly'}),
     (detector_entry, anomaly_detection.PrincipalComponentAnalysis, {'window_size': 15}),
     (detector_entry, anomaly_detection.PrincipalComponentAnalysis, {'window_size': 15, 'n_components': 0.5}),
     (detector_entry, anomaly_detection.KernelPrincipalComponentAnalysis, {'window_size': 15}),
@@ -429,6 +431,7 @@ class TestInterpretEntries:
     (detector_entry, anomaly_detection.LocalOutlierFactor),
     (detector_entry, anomaly_detection.MatrixProfileDetector),
     (detector_entry, anomaly_detection.MedianMethod),
+    (detector_entry, anomaly_detection.OneClassSupportVectorMachine),
     (detector_entry, anomaly_detection.PrincipalComponentAnalysis),
     (detector_entry, anomaly_detection.RobustPrincipalComponentAnalysis),
     # Preprocessors